extract 'State' from PhotoItem to MediaItem
