+++ src/gui/desktop/quick_views/media_item.cpp
@@ -1,7 +1,26 @@
 
 #include "media_item.hpp"
 
-void MediaItem::paint(QPainter* painter)
+
+MediaItem::MediaItem(QQuickItem* p)
+    : QQuickPaintedItem(p)
+    , m_state(State::NotFetched)
 {
 
+}
+
+
+MediaItem::State MediaItem::state() const
+{
+    return m_state;
+}
+
+
+void MediaItem::setState(MediaItem::State state)
+{
+    const bool changed = state != m_state;
+    m_state = state;
+
+    if (changed)
+        emit stateChanged();
 }
